[IMP] l10n_br,l10n_br_website_sale: better Brazilian addresses support
This improves the eCommerce address page for Brazil by replacing the
regular address fields with base_address_extended fields. l10n_br
depends on base_address_extended.

A downside is that the user has to select the city_id out of a large
list of cities. To make this easier, this also adds an automatic
selection of city_id and state_id based on zip code using the new
l10n_br_zip_ranges field on res.city. The data for this field was
taken directly from its source [1] to make it easy to update in the
future.

task-3860901

[1] https://github.com/datasets-br/city-codes