[FIX] web: snailmail global address spacing
The JP report is displaying the company name at the end of the address.
This scenario was not taken in consideration in the previous fix
odoo/odoo@9a206d858d66

This lead to rethinking the approach to a more pragmatic approach
targeting anything under the address block instead of being specific.

This will lead to 16 extra pixels on the 4 reports with the
o_snail_mail class, but provide a more consistent result and less
customization.

task-4076930