[IMP] point_of_sale, pos_*: use x_commands from frontend.
*: pos_sale, pos_loyalty, pos_restaurant, pos_event

Before all lines was sent to the backend to be processed, if the backend
had a line that wasn't sent by the frontend, it was removed from the
order.

Now the frontend sends only modified lines to the backend, if a line is
deleted from the frontend, it will be deleted from the backend too. But
the backend doesn't remove lines that weren't sent by the frontend.

This change was made to avoid losing information when the backend delete
a line that wasn't sent by the frontend.