[FIX] website_profile: paginate leaderboard properly when grouping
Problem:

When displaying the leaderboard grouped by month or week, users were shown in the wrong order. This issue occurred because we were always sorting by karma_points (global count), paginating, and then joining the scores of that week/month and sorting again, which is incorrect. Additionally, the current user was never shown due to this issue.

Steps to reproduce:

- Install the eLearning module.
- Create a few users with different karma_points (more than 25 to have 2 pages).
- Go to /profile/users.
- Group by week.
- Paginate, and you will notice that the order is wrong; the first user on the second page might have more points than users on the first page.
Also, when the logged-in user is not on that page, they do not appear at the bottom.

opw-3979785