[ADD] event_mail_scale: improve event mail scalability
Purpose of this module is to make event-based schedulers working using
iterative batch of registrations. Idea is to make batch of attendees
to contact, commit regularly, and continue where it has stopped if
it does not finish within a cron lifetime for some reason.

It requires to know where the scheduler did stop which is why this is
done in a separate module. Enterprise code is taken into account for
social notifications, which somehow breaks modular approach but is
simpler than having several bridge modules.

Task-3084943: Event: Improve communication scheduler scalability