[IMP] website: improve media icons
This commit enhances the social media icons by reducing their width and
height when the layout is set to disk or square with a small size.
Additionally, it removes the left and right margins of the icons when
the layout is set to none and the size is small.

task-4107178