[ADD] ir_module: Translation of knowledge
In the industry we have knowledge article that should be translated in
the language of the user installing it.

To do so the knowledge body is put in a template that is traducted
and then wrote in the knowledge article body.

We check that the body and article exist before writing.

task-4063061