[IMP] html_editor: add `onEditorReady` callback on editor
In Knowledge, the editor has to show a helper when the document is empty.
This helper aims to help people writing their first article and discover
the features of Knowledge (the templates, the article items, etc.).

When the user initially opens an article in Knowledge, we need to check
if the content of the editable zone is empty. To do that, one could
overwrite the `onEditorLoad` of the HTML field and add the instructions
in that function to show the helper if the document is empty.

Unforthunatly, this function is called when the editor is not yet attached
to an editable zone. `editable` is therefore `null` when calling the `onEditorLoad` function. As a result, we can not determine if the document
is initially empty or not.

To fix that issue, we will add a new callback function named `onEditorReady`
in the `config` object passed to the editor. The editor will then call that
function when the editor is ready and attached to the editable zone.

In Knowledge, we will then be able to provide a `onEditorReady` function
to decide if the helper should be shown or not.

task-4061283