[FIX] mail: respect 'find_or_create_partners' parameter
Mail template rendering allows to keep emails instead of always
creating partners. However when invoked through the composer the
optional parameter is not correctly propagated.

This has no impact on standard code as no call is done with
this parameter being set to False but we should respect it.

Task-3084943: Event: Improve communication scheduler scalability