[PERF] event_mail: use composer for mailing
Purpose of this commit is to use the composer in mass mailing mode
instead of relying on low-level "send_mail" method of template model
(either directly, either through Event.mail_attendee).

To keep backward compatible behavior, we do not enforce partner
creation in the composer, allowing to create mails with emails instead
of always using partners. This is done using a context key as it is
currently not an option available on composer model, even if supported
in code itself.

Task-3084943: Event: Improve communication scheduler scalability