[FIX] web: prevent cancel action deleting *2m fields
When leaving the settings page with unsaved changes, a dialog window
is prompted for the user to confirm. The cancel action calls a discard
followed by a save which is causing o2m and m2m fields to be unlinked.
This commit removes the save action since it should not be necessary
when trying to cancel your changes.

opw-4050990