[FIX] bus: reduce concurrency when fetching notifications
When no cursor is available to fetch bus notifications, websockets
retry up to 10 times to acquire one. However, the delay between each
attempt is too short, resulting in minimal chance of recovery when the
connection pool is full. This will eventually disconnect only to find
out no cursor is available to initialize the connection either.

This PR introduces a semaphore to avoid multiple greenlets trying to
get a cursor at the same time. 10% of the cursors are kept to process
incoming connection requests.