[FIX] im_livechat: trigger step data wrongly inserted
Since [1], the return type of the `/chatbot/trigger/step` route
has changed. However the call site was not adapted. As a result,
the chat bot fails when triggering a question selection step.

Steps to reproduce:
- Create a chat bot script with two question selections
- Enable it on the website
- Follow the steps, the second one results in a crash

This PR adapts the chat bot code to properly use the new mail
store syntax.

[1]: odoo#171585

opw-4130439

X-original-commit: abd6c4f2cff986133760f970a376b354c6b4a37c