[IMP] mail, web: emoji picker position in chatter
Before this commit:

In mobile view, when using the emoji picker on the chatter, it replaces the
native keyboard but it does not take its position. It's possible to scroll
outside the emoji picker.

After this commit:

In mobile view, chatter emoji picker works just like discuss's emoji picker and
replaces native keyboard and takes the right position.

task-id:3757047