[IMP] html_editor: make datetime field in dynamic placeholder acc to user's tz
How to reproduce:
=====================
- Create an email template, add datetime field using /dynamic command
- Set the datetime on a record
- Send an email

Specifications:
=====================
Currently, datetime field in the email always display utc.

After this commit:
=====================
When the user inputs a datetime field using the /dynamic command, it will
now be displayed according to the recipient's or sender's timezone, with
the timezone shown if the partner_id field is defined in the applied model.

Task-4011120