[IMP] html_editor: apply color to list item marker
Before this commit:

It was not possible to apply color to list item markers.

After this commit:

This commit ensures that if a list item is selected completely, applying color
to it will apply color to the list item itself making sure the list item marker
also gets the color.

Note: The color is applied to marker only when the entire list item is selected,
not if list item is only partially selected.

task-3281223