[FIX] websites_slides: Elearning Mobile Zoom in and out
Upon zooming in on the elearning slides view from the mobile by finger pinch, the whole page zooms in and not just the content, which causes the content quality to be still blurry.

To Reproduce on Runbot:
1. Go to the eLearning module.
2. Go to "Edit" on any of the courses
3. Add a pdf content
4. From the phone, go to the webiste elearning for the same course
5. Click on the pdf content we uploaded
6. Try finger ping zooming in, the document get still blurry

The issue seems to be that there's no finger pinch zoom in and out, and when you try doing that the whole page zooms in and out, not changing the quality of the document.

So, the solution was to add the zoom in and out icon on small screen devices as well. But, upon adding those icons, the left side of the PDFViewerNav seems congested, so put the
zoom in and out icon to the right side besides the fullscreen icon.

opw-3893275