[FIX] point_of_sale: prevent error when clicking on order
When we try to add any product for order in the first browser and simultaneously
close the session in the second browser, the error will occur in the first
browser when we click on order.

Steps to reproduce:
- Install the 'point_of_sale' module(without demo data)
- Open the 'bar', copy the URL, and Open a POS session on 2 different browsers
- Now go to the second browser and then close the session there
- Now come to the first browser and add any product
- Click on 'Order'

Traceback: ValueError
Expected singleton: pos.config()

This error occurred at [1] where we didn't find 'config_id' in order.

This commit will fix the error by passing a session ID into the order, which was
before not found in order.

[1]- https://github.com/odoo/odoo/blob/42acdb4bffa5295ec834fc9d6258540683ab90ce/addons/pos_online_payment/models/pos_order.py#L16

sentry-5699019259