[IMP] mail: add a static method to inline editor content
This commit updates the HtmlMailField Component from the mail addon:

We put the part responsible for converting the editor content to an Html
element compatible with email clients (inline CSS style) into a static
method.
Thus, the same code can be reused in the member invite component of
the documents permission panel.

task-4014624