[IMP] hw_drivers: protect public routes with payload signature
Most routes on the IoT Box are public. We implemented a decorator to require a payload signature based on the token shared on connection between IoT Box and DB.
We updated this token to be a 16 bytes hex token as well.

As Odoo's hmac methods are meant to use the `database.secret` from `ir.config_parameter`, we had to re implement those methods to use our token.

The signatures are time based and have a window of 5 seconds to be validated.

Additionally, we only kept the token in the database generated url to provide the IoT Box in order to reduce its size:
`db_uuid` and `enterprise_code` are now provided by the db itself on IoT Box setup.

Enterprise PR: [https://github.com/odoo/enterprise/pull/69426](https://github.com/odoo/enterprise/pull/69426)
Task: 3082226