[FIX] mrp: document in bom chatter not showing
In https://github.com/odoo/odoo/pull/152286, we changed the documents
that are added to the chatter of a BoM to be added to the product_product directly
(or the template if there are no product on the BoM).

However, some changes in mail made the just-added document not appear untill reload.
This commit fixes it by reloading the chatter when adding an attachment.