[FIX] mail: less spacing between author and msg body of reply
When a message is a reply, the replied message preview had too
much spacing between author and message boby. This comes from
earlier design that had weak stylistic difference of author and
message, so it relied on extra spacing.

With recent new design, the author is bold and quite easy to see,
so we can have less spacing, which looks better and also prevents
2 characters from being truncated!