[FIX]  TO REMOVE html_editor: stageSelection fail with mutation in currentStep
Before this commit, stageSelection could update the currentStep selection
with mutations (‘add’, ‘remove’ and ‘characterData’) in progress.
We should not have mutations of this type without adding them to a step.
This type of mutation can cause an invalid selection to be generated when
we remove it. We should therefore never have this type of situation.

We shouldn't have any ‘attributes’ mutations either. Unfortunately,
this is too common at the moment, so we've chosen to ignore them in this commit.
Also, it doesn't cause selection problems, so it's less critical.

In order to avoid silent errors, we will throw an error when this situation
occurs in order to detect mutations that should be added to a step or removed from the history.

This commit will therefore delete two of them.
The first occurs when an overlay is opened, and the history must ignore these mutations.
The second occurred when the table picker was opened, and an unnecessary node was added.