[FIX] web: undefined selection data
This commit fixes an error that sometimes happen on selection fields
when there are no available options, it throws an undefined/null error.
This is caused by how the code tries to access the `[1]` immediately
without checking if there exist any item to begin with.

This happens because of the new dynamic selection field implemented in
the previous commit. It is now possible for an `options` object to be
empty due to how we design the field. We want to avoid traceback even
when the selection field is empty.

task-id: 3358316