[ADD] l10n_gr_mydata: Send Classification to MyDATA
MyDATA is a platform created by Greece's tax authority, The Independent
Authority for Public Revenue (IAPR), to digitize business tax and
accounting information declaration. All companies performing accounting,
according to Greek accounting standards, will be required to transmit
the information to the myDATA platform beginning in July of 2021.

- Customer Invoices

Customer Invoice form for Greece companies will now have a few fields
required for specifying classification data to be send to MyDATA.

In the Send & Print Wizard, a checkbox "Send to MyDATA"is added, and
when activated, Odoo will generate an XML of the selected invoice(s) and
send them to the MyDATA API.

Any errors detected before and after the send will then be saved in a
document object, displayed in the 'myDATA' tab in the move form view.

- Vendor Bills

In Greece, one's expense is another's invoice. When reporting to MyDATA
for our expense classification, we need to fetch the invoice(s) issue
by third-party issuer, fill the classification needed and send them
back to MyDATA.

To implement this behavior, we created a daily cron to fetch xml data
from MyDATA related to the designated Greece companies, and create draft
vendor bills based on that data. The user can then fill the myDATA
classification data needed and send them back to MyDATA via a button
in vendor bills view form after confirming the bill.

task-id: 3358316