[FIX] web_tour: fixed tour traceback in last step if is not active and warn step
In all the tours containing last step as step with function or run(warn step)
there is a trace-back because every warn step is same as not active step.
Thus the subsequent logic for non active step is getting executed and In there
the currentActionIndex is being increased above this.actions.length. This all
happens after tour is ended. So we get currentAction as null and the program
tries to setListeners thus causing the TypeError Traceback.

task- 4009134