[FIX] website: Fix form snippet field options and styling loss on undo
In a form snippet, enabling the 'Description' option for a field and
then undoing this action caused a loss of options and styling for that
field.

The issue was resolved by removing a redundant line of code.

This line was adding an extra step in the mutation history (list of
changes).
Due to this additional step, when undoing the action, the classes from
the field were also removed. These removed classes were the reason for
the loss of styling and options.

task-4013551