[IMP] stock: fetch rule from location in one step
Currently, when trying to find a pull rule for a specific location, we
try to find a different a rule using a few different filters, then if
none are found, we try the whole process again from the parent location
until we find a rule or there are no parent location anymore.

Instead of doing that, we could fetch the rules that have their location
being a parent of the searched rule. This avoids avoids doing the whole
process again.