[FIX] stock_account: use the correct accounts for adjustment & scrap
This commit fixes the automatic valuation of stock adjustments and
scraps stock moves. The valuation logic will be as follows: the
expense account will be debited in case of negative
adjustment or scraps (if an outgoing account is specified on
the adjustment or scrap location, it will be used instead for each
scenario respectively). The category stock valuation account will be
credited in both cases. In case of positive adjustment, the general
income account will be credited (or the location incoming valuation
account if defined) and the stock valuation account will be
debited.

Task-4013809