[ADD] product_warranty: add module for warranty
In Odoo, there is not enough to help users managing the warranties,
maintenance and tracked serial numbers. 
This commit adds a new product warranty module to track the warranty of
delivered lots.

This will centralize the link with Tasks and serial numbers for history
overview by adding a specific view on the customer.

task-4100027