[IMP] event: prepare scheduler scalability
There is an ongoing work to make event mail schedulers more scalable. This is
currently done as an internal odoo development that is going to land in master
soonish. In order to be ready for the freeze and be sure to have the feature
for the upcoming v18 a necessary field is already added.

It will be used to make event-based communication iterative, and for that purpose
we need to know the last contacted attendee, to start again at that point at
next iteration.

Task-