[IMP] stock: access property supplier/customer through getters
While the properties `property_stock_customer`/`property_stock_supplier`
work well for regular use, they're not ideal for inter-company use
cases, as creating a property from each company for each other company
partners (and potentially their sub-partners) will quickly create a huge
amounts of `ir.property`, all pointing to the same location.

To avoid that, we remove the inter-company transit location from the
`property_stock_customer`/`property_stock_supplier` and instead set it
in a new property when inter-company transactions are enabled.

This means that we need to be able to override the access to those
properties if we're in a inter-company context, thus the change to a
getter.