[IMP] html_editor, project: use the HtmlViewer in the history dialog
Using the HtmlViewer in the History Dialog will allow to mount embedded
components in the comparison previews. This will offer a far better
representation of the state of the html_field and its changes.

History Dialog assets are moved to the `html_editor` module to remove them from
the `components` folder of `web_editor`, which bundled the History Dialog into
more assets than necessary.

task-3672747

Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>