[IMP] html_editor: share method used for the comments plugin in Knowledge
The comments plugin in knowledge uses unremovable protected empty `<a>` nodes as
"beacon" to "wrap" DOM positions related to a comment.

This commit shares some low-level functions so that the comment plugin can
handle everything by itself (since the nodes are empty and protected, the
current plugins are impeded and the implemented logic does not directly work).

In the future, these functions have to be un-shared again, but it will need a
bit of refactoring so that the plugins can also handle the Knowledge use case.

Also related to the comments feature in Knowledge, `fetchMessages` of the
`thread_model` is split to allow overrides. In Knowledge, there are multiple
threads with multiple messages to be loaded at once, and it will be done in
batches.

task-3672747

Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>