[FIX] html_editor: handle full table selection properly
- When there is a text above and below a table and user
  starts selecting the content from the text above the
  table using keyboard, he needs to press shift + ArrowDown
  multiple times to reach the text below table.

- Similar scenario occurs in case of selecting from text
  below to the text above.

- This happens because when selection is inside the table
  as well as outside of it, table gets fully selected through
  customized selection but the actual focusNode is still the
  first cell of the table.

Desired behaviour after commit:

Now when selection enters the table from outside, last cell
of the table cell is set as focusNode (first cell in case
of inverse selection) in order to select the full table.
Due to this, user can select the content after the table
in just one press.

task-3442805