[IMP] html_editor: add the excalidraw embedded component
This commits moves the `draw_behavior` from knowledge to html_editor.
It is a simple component with an iframe containing an excalidraw URL.
It brings the ability to display interactive drawing boards in any HtmlField
(i.e. a task mockup) so that the user does not need to leave odoo to consult
them.

task-3672747

Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>