[FIX] account: audit_trail activated without any companies with it
So, we have 2 things we solve, coming from:
Go to Accounting/Invoicing, Customers and create a new Customer.
Delete it.
You get an error talking about audit trail while you don't have
it activated.

We now only check if we should block the deletion of a message of
a partner if this partner has at least one journal entry with a
company that has audit trail activated.