[FIX] web_editor: enable the use of "|" in form field
Before this commit, after selecting a field containing the char "|" for
conditional display, the field on which the condition is set will never
appear again.

This commit enables the use of any characters in form fields.

Steps to reproduce the bug:
- Add a form
- Add two fields (A and B)
- Rename the field A with a string that contains "|"
- Set the field B visibility to "Visible only if"
- Set the field A as the visibility condition for field B
(field B visible only if field A contains 'hello', for example)
- Save the changes
- Write an input in the field A according to the visibility condition
(The second field does not appear)

task-3893749