[FIX] html_editor: don't display a peer selection inside a protected element
Don't display a peer selection if it is inside a protected element (like an
embedded view). This could be improved in the future, but currently since
protected elements are not editable, the selection is fixed as a range to fully
contain the non-editable element, and this is not what we want.

task-3672747

Co-authored-by: Damien Abeloos <abd@odoo.com>
Co-authored-by: Adrien Schoffeniels <adsc@odoo.com>
Co-authored-by: Julien Banken <jbn@odoo.com>
Co-authored-by: Thomas Josse <thjo@odoo.com>