[FIX] account, account_payment: portal payment and invoice status
Fix the status of invoices when registering a payment. Also improves the portal view
depending on the invoice payment state.

The following cases assume that no outstanding account is set on the bank journal

Portal:
 - In case the invoice payment state is 'in payment', the portal should display "Processing payment",
 the "Next installment" section should be hidden, and a banner should appear in the popup of
 the payment (when pressing the "Pay Now" button).

Portal list view:
- Same payment state should be displayed as in the portal form view, e.g. "Processing Payment"

Invoice Payment state:
 - If a payment is created for an invoice, then the state of the invoice is set to 'in payment'
When the invoice is reconciled, and the payment is manually validated (This manual operation
of validating the payment will be made automatically in another task)
if the payment corresponds to a portion of the invoice amount, the invoice payment state is 'Partial'
if the payment corresponds to the full invoice amount, the invoice payment state is 'Paid'
If the payment is rejected, then the invoice payment state is reset
If the payment is confirmed (before any reco), then the invoice payment state keeps being "in payment"

 task-4212954