[REF] html_editor: introduce delegateTo
There is a common pattern in the editor plugins where we allow other
plugins to take over the execution of a command. In order to make it
explicit in the code, this commit introduces the `delegateTo` plugin
method, that will be used to call these overriding functions, registered
in resources postfixed with "_overrides".

task-4266746