[REF] html_editor: remove dispatch
The goal of this commit is remove the use of `Plugin.dispatch` and
`Plugin.handle`.

*********************
Better explain this:
- dispatch used for system commands -> call shared method instead
- dispatch for a user command -> can registered command via execCommand
- dispatch as an event (multiple handlers) -> register handlers as
resources, and call each of them.  A plugin method "dispatchTo" is
introduced (or does it replace dispatch?) to abstract this and better
convey the idea of an event.
****

Reasons:

1) It is a redundant mechanism, we can achieve the same result through the use
of shared and resources.
2) Using the shared or the resources is more explicit. If we need to call a
command from a specific plugin, we should depend on the plugin and use the
shared. If we need multiple plugins to react to an "event", we should use the
resources.
3) We should distinguish between a "user command" and a "system command" or
"system event".
- A "user command" is anything that a user could configure as a shortcut,
toolbar item, powerbox item, or power button.
- A "system command" is a shared method.
- A "system event" is a resource.
4) When debugging, it was harder to "step into" a dispatched command as we would
step into the `handleCommand` of every plugins.

The resources `toolbarItems` and `powerboxItems` now inherit the properties of a
user command through the `commandId` property (if specified).

The resource `powerButtons` now use the definition of a user command instead
of a powerbox item.

The `toolbarItem` `Component` does not includes dispatch in the props by default
anymore, we need to specify the callbacks in the props of the `toolbarItem`
explicitly.

******* TODO: ****
move this up, after the reasons list.

In the editor, the pattern to dispatch events from one plugin to other
plugins is to use the resource mechanism that aggregates callbacks to
a specific resource. In order to make it clear when reading code that
we want to dispatch an event, this commit introduces the Plugin method
`dispatchTo` that communicates the intent to the reader. The reffered
callbacks are posfixed with `_handlers`.

task-4266746

label -> title

rm hidden dependencies with userCommand.run

remove use of execCommand / run