[IMP] html_editor: rename Image command to Media
The "/Image" powerbox command opens a dialog that allows the user to
insert images, documents and icons.

This commit updates the name and description of the "Image" powerbox
command in order to reflect its actual capabilities.

The new name "Media" is more genereric and reflects better the dialog
title ("Select a media").

The new description ("Insert image, document or icon") matches the names
of that tabs in the dialog.

This commit also makes sure that /image, /document and /icon all result
in the same Media powerbox command.